+++ consensus/validator.go
@@ -613,7 +613,7 @@ func (cv *ConsensusValidator) ProcessNotaryAnnounceMessage(notaryMsg *NotaryAnno
 
 	// XXX: Start pacemaker here at this time.
 	// TODO: we should design a meessage like committee decided and ask every committee member start pacemaker
-	cv.csReactor.csPacemaker.Start(cv.csReactor.curHeight, 0)
+	cv.csReactor.csPacemaker.Start(uint64(cv.csReactor.curHeight), 0)
 	return true
 }
